chore: intro window design changes
